+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- METRICS: station_clicks, api_calls
+- API: codec parameter to advanced search
+
+### Fixed
+- Use utc timestamp for clicks in database
 
 ## [0.6.11] 2020-02-17
 ### Added